Changed the initial index file using VS Code
I added a new meta which just includes that I am the author. I also changed the title to add a name and job position. The final change is adding a second line of text by wrapping the lines of code in <p> which separates the sentences onto separate lines.